QBrain/rag_lab/pyproject.toml
[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[project]
name = "qbrain-rag"
version = "0.1.0"
description = "QBrain RAG mini-lab (research / thesis experiments)"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.0",
  "pandas>=2.0.0",
  "numpy>=1.26.0",
  "scikit-learn>=1.3.0",
  "openai>=1.40.0",
  "langchain>=0.3.0",
  "langchain-openai>=0.2.0",
  "langchain-community>=0.3.0",
  "langchain-core>=0.3.0",
  "langchain-text-splitters>=0.3.0",
  "pymupdf>=1.24.0",
  "faiss-cpu>=1.8.0",
  "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = ["jupyter>=1.0.0", "notebook>=7.0.0"]

[tool.setuptools.packages.find]
where = ["src"]
